Добавил костыль ограничение на скачивание

+++ src/main/java/com/github/yourbootloader/config/YDProperties.java
@@ -11,7 +11,7 @@ import java.nio.file.Paths;
 public class YDProperties {
     private String downloadPath;
 
-    private Long maxFileSize;
+    private Integer maxFileSize;
 
     public Path getDownloadPath() {
         return Paths.get(downloadPath);
